# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AbsencesRegisteredResponse": {
        "properties": {
          "absenceIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "absenceIds"
        ],
        "type": "object"
      },
...
      "AnimalsRegisteredResponse": {
        "properties": {
          "animalIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "animalIds"
        ],
        "type": "object"
      },
...
      "EndStayRequest": {
        "properties": {
          "on": {
            "type": "string"
          }
        },
        "required": [
          "on"
        ],
        "type": "object"
      },
      "EntranceCreatedResponse": {
        "properties": {
          "condominiumId": {
            "format": "uuid",
            "type": "string"
          },
          "entranceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "condominiumId",
          "entranceId"
        ],
        "type": "object"
      },
      "EntranceView": {
        "properties": {
          "condominiumId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "managementForm": {
            "type": "string"
          }
        },
        "required": [
          "condominiumId",
          "id",
          "label",
          "managementForm"
        ],
        "type": "object"
      },
...
      "HouseholdRegisteredResponse": {
        "properties": {
          "memberIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "memberIds"
        ],
        "type": "object"
      },
...
      "NewAbsenceRequest": {
        "properties": {
          "absentFrom": {
            "type": "string"
          },
          "absentTo": {
            "type": "string"
          }
        },
        "required": [
          "absentFrom",
          "absentTo"
        ],
        "type": "object"
      },
      "NewAnimalRequest": {
        "properties": {
          "species": {
            "type": "string"
          },
          "validFrom": {
            "type": "string"
          },
          "vetPassportNo": {
            "type": "string"
          }
        },
        "required": [
          "species"
        ],
        "type": "object"
      },
      "NewMemberRequest": {
        "properties": {
          "isChildUnder6": {
            "type": "boolean"
          },
          "validFrom": {
            "type": "string"
          }
        },
        "required": [
          "isChildUnder6"
        ],
        "type": "object"
      },
      "NewUnitRequest": {
        "properties": {
          "areaM2": {
            "type": "number"
          },
          "designation": {
            "type": "string"
          },
          "idealParts": {
            "type": "string"
          },
          "separateEntrance": {
            "type": "boolean"
          },
          "unitType": {
            "type": "string"
          }
        },
        "required": [
          "designation",
          "idealParts",
          "separateEntrance",
          "unitType"
        ],
        "type": "object"
      },
...
      "RegisterAbsencesRequest": {
        "properties": {
          "absences": {
            "items": {
              "$ref": "#/components/schemas/NewAbsenceRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "absences"
        ],
        "type": "object"
      },
      "RegisterAnimalsRequest": {
        "properties": {
          "animals": {
            "items": {
              "$ref": "#/components/schemas/NewAnimalRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "animals"
        ],
        "type": "object"
      },
      "RegisterEntranceRequest": {
        "properties": {
          "address": {
            "type": "string"
          },
          "condominiumId": {
            "format": "uuid",
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "managementForm": {
            "type": "string"
          }
        },
        "required": [
          "label",
          "managementForm"
        ],
        "type": "object"
      },
...
      "RegisterHouseholdRequest": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/NewMemberRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "members"
        ],
        "type": "object"
      },
...
      "RegisterUnitsRequest": {
        "properties": {
          "units": {
            "items": {
              "$ref": "#/components/schemas/NewUnitRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "units"
        ],
        "type": "object"
      },
...
      "StayEnded": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "validFrom": {
            "format": "date",
            "type": "string"
          },
          "validTo": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "validFrom",
          "validTo"
        ],
        "type": "object"
      },
...
      "UnitView": {
        "properties": {
          "areaM2": {
            "type": "number"
          },
          "designation": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "idealPartsPct": {
            "type": "number"
          },
          "separateEntrance": {
            "type": "boolean"
          },
          "unitType": {
            "type": "string"
          }
        },
        "required": [
          "designation",
          "id",
          "idealPartsPct",
          "separateEntrance",
          "unitType"
        ],
        "type": "object"
      },
      "UnitsCreatedResponse": {
        "properties": {
          "unitIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "unitIds"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Generated from the running code (ADR-013): it states what runs, nothing more. Each operation names the rules it serves in `x-rules`. Operations designed but not built yet are listed in `tools/build_openapi.py` (PLANNED), not here.",
    "summary": "Управление на етажна собственост по ЗУЕС",
    "title": "domuvai api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/registry/entrances": {
      "get": {
        "operationId": "get_registry_entrances",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntranceView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "List the registered entrances",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-ORG-001"
        ]
      },
      "post": {
        "operationId": "post_registry_entrances",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEntranceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntranceCreatedResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Register an entrance — founding its building at an address, or joining an existing building",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-ORG-001"
        ]
      }
    },
...
    "/api/registry/entrances/{entranceId}/units": {
      "get": {
        "operationId": "get_registry_entrances_entranceId_units",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnitView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "List the units of an entrance",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-ORG-001",
          "PM-ORG-002"
        ]
      },
      "post": {
        "operationId": "post_registry_entrances_entranceId_units",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUnitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitsCreatedResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Add units with their ideal parts",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-ORG-002",
          "PM-ORG-009"
        ]
      }
    },
    "/api/registry/entrances/{entranceId}/units/{unitId}/absences": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_units_unitId_absences",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAbsencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbsencesRegisteredResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "File an absence declaration — the long-absence exemption",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-FEE-006",
          "PM-FEE-007"
        ]
      }
    },
    "/api/registry/entrances/{entranceId}/units/{unitId}/animals": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_units_unitId_animals",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAnimalsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalsRegisteredResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Record the animals kept in a unit",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-BOOK-005",
          "PM-FEE-009"
        ]
      }
    },
    "/api/registry/entrances/{entranceId}/units/{unitId}/animals/{animalId}/end": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_units_unitId_animals_animalId_end",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "animalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndStayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StayEnded"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Record or correct the day an animal left the unit",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-BOOK-005",
          "PM-BOOK-008"
        ]
      }
    },
    "/api/registry/entrances/{entranceId}/units/{unitId}/household": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_units_unitId_household",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdRegisteredResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Register a unit's household — who counts for per-person charges",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-FEE-005",
          "PM-FEE-008"
        ]
      }
    },
    "/api/registry/entrances/{entranceId}/units/{unitId}/household/{memberId}/end": {
      "post": {
        "operationId": "post_registry_entrances_entranceId_units_unitId_household_memberId_end",
        "parameters": [
          {
            "in": "path",
            "name": "entranceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "unitId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "memberId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndStayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StayEnded"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "summary": "Record or correct the day a resident left — their occupancy range closes",
        "tags": [
          "registry"
        ],
        "x-rules": [
          "PM-BOOK-008"
        ]
      }
    },
...
  },
...
}